Extract OpenGL frame pipeline into dedicated backend and slim ImGuiController to UI/input responsibilities.
