+++ src/Tunnerer.Desktop/Rendering/RenderBackendFactory.cs
@@ -24,8 +24,9 @@ public static unsafe class RenderBackendFactory
     private static RenderBackendServices CreateOpenGlServices(Sdl sdl, Window* window, int windowW, int windowH)
     {
         var gl = GL.GetApi(name => (nint)sdl.GLGetProcAddress(name));
+        var imgui = new ImGuiController(gl, sdl, window, windowW, windowH);
         return new RenderBackendServices(
-            Backend: new ImGuiController(gl, sdl, window, windowW, windowH),
+            Backend: new OpenGlGameRenderBackend(gl, imgui),
             Textures: new TextureManager(gl));
     }
 }
